+++ backend/src/controllers/profileController.ts
@@ -8,6 +8,7 @@ import fs from "fs";
 import { unlink } from "fs/promises";
 import { v4 as uuidv4 } from "uuid";
 import validator from "validator";
+import jwt from "jsonwebtoken";
 
 export const ProfileController: IController = {
   // GET ALL PROFILES
@@ -70,6 +71,39 @@ export const ProfileController: IController = {
     }
   },
 
+  getAuthenticatedUserProfile: async (req: Request, res: Response) => {
+    try {
+      const token = req.cookies.jwt;
+      console.log({ token });
+      const decodedToken = jwt.verify(
+        token,
+        process.env.TOKEN as string
+      ) as jwt.JwtPayload;
+      const id = decodedToken.userId as string;
+      const myProfile = await dataSource
+        .getRepository(User)
+        .createQueryBuilder("user")
+        .select([
+          "user.id",
+          "user.username",
+          "user.image",
+          "user.role",
+          "user.city",
+          "user.email",
+        ])
+        .leftJoinAndSelect("user.createdPoi", "createdPoi")
+        .where("user.id = :id", { id })
+        .getOne();
+      if (myProfile === null) {
+        return res.status(404).send("User not found");
+      } else {
+        return res.status(200).send(myProfile);
+      }
+    } catch (err) {
+      return res.status(400).send({ error: "Error while reading user" });
+    }
+  },
+
   // UPDATE PROFILE BY ID
 
   updateProfile: async (req: Request, res: Response): Promise<void> => {
@@ -87,20 +121,23 @@ export const ProfileController: IController = {
           res.status(400).send({
             error: `Field must contains only characters`,
           });
-          await unlink(`./public/user/${req.file?.filename}`);
+          if (req.file !== undefined)
+            await unlink(`./public/user/${req.file?.filename}`);
+          return;
         }
       };
 
       const inputString: string[] = [username, email, city, password];
-      inputString.forEach((value) => {
-        if (value) checkIfStringAndNotEmpty(value);
+      inputString.forEach(async (value) => {
+        if (value !== null) await checkIfStringAndNotEmpty(value);
       });
 
       // check enum in role
       const roles: UserRole[] = Object.values(UserRole);
-      if (role && !roles.includes(role)) {
+      if (role !== null && !roles.includes(role)) {
         res.status(400).send({ error: "User role does not exist" });
-        await unlink(`./public/user/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/user/${req.file?.filename}`);
         return;
       }
 
@@ -110,7 +147,8 @@ export const ProfileController: IController = {
         .findOneBy({ id });
       if (profileToUpdate === null) {
         res.status(404).send({ error: "User not found" });
-        await unlink(`./public/user/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/user/${req.file?.filename}`);
 
         return;
       }
@@ -120,37 +158,39 @@ export const ProfileController: IController = {
         .getRepository(User)
         .findOne({ where: { id: userId } });
 
-      if (role && currentUser?.role !== UserRole.ADMIN) {
+      if (role !== null && currentUser?.role !== UserRole.ADMIN) {
         res.status(403).send({
           error: "You are not authorized to update this property",
         });
-        await unlink(`./public/user/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/user/${req.file?.filename}`);
 
         return;
       }
 
       // Check if connected user has the same id than profile to update or if he is admin
-
       if (
-        currentUser?.id !== profileToUpdate.id ||
+        currentUser?.id !== profileToUpdate.id &&
         currentUser?.role !== UserRole.ADMIN
       ) {
         res.status(403).send({
           error: "You are not authorized to update this profile",
         });
-        await unlink(`./public/user/${req.file?.filename}`);
+        if (req.file !== undefined)
+          await unlink(`./public/user/${req.file?.filename}`);
 
         return;
       }
 
       // check if username already exist in db
-      if (username) {
+      if (username !== null) {
         const usernameAlreadyExist = await dataSource
           .getRepository(User)
           .count({ where: { username } });
         if (usernameAlreadyExist > 0) {
           res.status(409).send({ error: "Username already exists" });
-          await unlink(`./public/user/${req.file?.filename}`);
+          if (req.file !== undefined)
+            await unlink(`./public/user/${req.file?.filename}`);
         }
         if (
           !validator.matches(
@@ -161,27 +201,30 @@ export const ProfileController: IController = {
           res.status(401).send({
             error: "Username must contain 3 to 20 characters and no symbol",
           });
-          await unlink(`./public/user/${req.file?.filename}`);
+          if (req.file !== undefined)
+            await unlink(`./public/user/${req.file?.filename}`);
         }
       }
 
       // check if email already exist in db
-      if (email) {
+      if (email !== null) {
         const emailAlreadyExist = await dataSource
           .getRepository(User)
           .count({ where: { email } });
         if (emailAlreadyExist > 0) {
           res.status(409).send({ error: "Email already exists" });
-          await unlink(`./public/user/${req.file?.filename}`);
+          if (req.file !== undefined)
+            await unlink(`./public/user/${req.file?.filename}`);
         }
         if (!validator.isEmail(email)) {
           res.status(401).send({ error: "Incorrect email format" });
-          await unlink(`./public/user/${req.file?.filename}`);
+          if (req.file !== undefined)
+            await unlink(`./public/user/${req.file?.filename}`);
         }
       }
 
       // rename the new file and delete the older one
-      if (req.file) {
+      if (req.file !== undefined) {
         // rename
         const originalname = req.file.originalname;
         const filename = req.file.filename;
@@ -190,7 +233,7 @@ export const ProfileController: IController = {
           `./public/user/${filename}`,
           `./public/user/${newName}`,
           (err) => {
-            if (err) throw err;
+            if (err !== null) throw err;
           }
         );
         req.body.image = `/public/user/${newName}`;
@@ -205,7 +248,8 @@ export const ProfileController: IController = {
       res.status(200).send("Updated user");
     } catch (err) {
       res.status(400).send({ error: "Error while updating user" });
-      await unlink(`./public/user/${req.file?.filename}`);
+      if (req.file !== undefined)
+        await unlink(`./public/user/${req.file?.filename}`);
     }
   },
 
@@ -229,7 +273,7 @@ export const ProfileController: IController = {
       }
 
       if (
-        currentUser?.id !== profileToDelete.id ||
+        currentUser?.id !== profileToDelete.id &&
         currentUser?.role !== UserRole.ADMIN
       ) {
         res.status(403).send({
@@ -271,7 +315,7 @@ export const ProfileController: IController = {
         .findOne({ where: { id: userId } });
 
       if (
-        currentUser?.id !== userToUpdate.id ||
+        currentUser?.id !== userToUpdate.id &&
         currentUser?.role !== UserRole.ADMIN
       ) {
         res.status(403).send({
@@ -324,7 +368,7 @@ export const ProfileController: IController = {
         .findOne({ where: { id: userId } });
 
       if (
-        currentUser?.id !== userToUpdate.id ||
+        currentUser?.id !== userToUpdate.id &&
         currentUser?.role !== UserRole.ADMIN
       ) {
         res.status(403).send({
@@ -385,7 +429,7 @@ export const ProfileController: IController = {
         .findOne({ where: { id: userId } });
 
       if (
-        currentUser?.id !== userToUpdate.id ||
+        currentUser?.id !== userToUpdate.id &&
         currentUser?.role !== UserRole.ADMIN
       ) {
         res.status(403).send({
@@ -441,7 +485,7 @@ export const ProfileController: IController = {
         .findOne({ where: { id: userId } });
 
       if (
-        currentUser?.id !== userToUpdate.id ||
+        currentUser?.id !== userToUpdate.id &&
         currentUser?.role !== UserRole.ADMIN
       ) {
         res.status(403).send({
